pyproject.toml
[project]
name = "bioproj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "biopython>=1.85",
    "chromadb>=1.5.7",
    "fastapi>=0.128.8",
    "groq>=1.0.0",
    "numpy>=2.0.2",
    "ollama>=0.6.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.39.0",
]
